@@ -127,7 +127,7 @@ public class JobSchedulerPlugin extends Plugin implements ActionPlugin, Extensib
         clusterService.addListener(this.sweeper);
         clusterService.addLifecycleListener(this.sweeper);
 
-        return Collections.emptyList();
+        return List.of(this.lockService);
     }
 
     @Override
